Rework SSRNode, add serializable JSX
