+++ packages/dom-expressions/src/serializer.js
@@ -1,4 +1,5 @@
-import { Feature, GLOBAL_CONTEXT_API_SCRIPT, Serializer, getCrossReferenceHeader } from "seroval"
+import { Feature, GLOBAL_CONTEXT_API_SCRIPT, Serializer, createPlugin, getCrossReferenceHeader } from "seroval"
+import { SSRNode } from "./ssr-node";
 
 const ES2017FLAG =
   Feature.AggregateError // ES2021
@@ -6,6 +7,27 @@ const ES2017FLAG =
   | Feature.BigIntTypedArray // ES2020;
 
 const GLOBAL_IDENTIFIER = '_$HY.r'; // TODO this is a pending name
+
+const JSXPlugin = createPlugin({
+  tag: 'SolidJSX',
+  test(value) {
+    return value.constructor === SSRNode;
+  },
+  parse: {
+    sync(value, ctx) {
+      return ctx.parse(value.t);
+    },
+    async(value, ctx) {
+      return ctx.parse(value.t);
+    },
+    stream(value, ctx) {
+      return ctx.parse(value.t);
+    },
+  },
+  serialize(node, ctx) {
+    return '_$HY.tmpl(' + ctx.serialize(node) + ')';
+  },
+});
 
 export function createSerializer({ onData, onDone, scopeId, onError }) {
   return new Serializer({
@@ -15,6 +37,9 @@ export function createSerializer({ onData, onDone, scopeId, onError }) {
     onData,
     onDone,
     onError,
+    plugins: [
+      JSXPlugin,
+    ],
   });
 }
