+++ packages/dom-expressions/src/ssr-node.js
@@ -1,0 +1,23 @@
+export class SSRNode {
+  constructor(t) {
+    this.t = t;
+  }
+}
+
+export function resolveSSRNode(node, top) {
+  const t = typeof node;
+  if (t === "string") return node;
+  if (node == null || t === "boolean") return "";
+  if (Array.isArray(node)) {
+    let prev = {};
+    let mapped = "";
+    for (let i = 0, len = node.length; i < len; i++) {
+      if (!top && typeof prev !== "object" && typeof node[i] !== "object") mapped += `<!--!$-->`;
+      mapped += resolveSSRNode((prev = node[i]));
+    }
+    return mapped;
+  }
+  if (t === "object") return node.t;
+  if (t === "function") return resolveSSRNode(node());
+  return String(node);
+}
